extract Signature module from Rekord
