+++ lib/rubygems/sigstore/rekor.rb
@@ -4,3 +4,4 @@ end
 require "rubygems/sigstore/rekor/api"
 require "rubygems/sigstore/rekor/log_entry"
 require "rubygems/sigstore/rekor/rekord"
+require "rubygems/sigstore/rekor/signature"
