+++ lib/rubygems/sigstore/rekor/rekord.rb
@@ -1,34 +1,6 @@
-require "rubygems/sigstore/cert_chain"
 require "rubygems/sigstore/rekor/log_entry"
+require "rubygems/sigstore/rekor/signature"
 
 class Gem::Sigstore::Rekor::Rekord < Gem::Sigstore::Rekor::LogEntry
-  def signature
-    @signature ||= begin
-      signature = Base64.decode64(body.dig("spec", "signature", "content"))
-      raise "Expecting a signature in #{body}" unless signature
-      signature
-    end
-  end
-
-  def cert_chain
-    Gem::Sigstore::CertChain.new(cert)
-  end
-
-  def signer_email
-    cert_chain.signing_cert.subject_alt_name
-  end
-
-  def signer_public_key
-    cert_chain.signing_cert.public_key
-  end
-
-  private
-
-  def cert
-    @cert ||= begin
-      cert = Base64.decode64(body.dig("spec", "signature", "publicKey", "content"))
-      raise "Expecting a publicKey in #{body}" unless cert
-      cert
-    end
-  end
+  include Gem::Sigstore::Rekor::Signature
 end
